Add Quick events support in Quickspace
This adds a new UI to Quickspace when events are available such as Now
Playing, Alarms, or whatever we configure as an event. This logic is
inspired by Google's event infrastructure in Smartspace but completely
written from scratch to be housed in Quickspace.

Initial events are:

- Device intro: Shows on first boot and setup
- Ambient Play (Now Playing): Shows what song is playing and tapping
the song will launch the results in Google search

Change-Id: I2742c13043b816c54bc2cefdfd415f52f2e76d3f
Signed-off-by: spezi77 <spezi7713@gmx.net>